[CVAT-UI]: Appearance settings in AAM, keyframe navigation and other buttons in AAM (#1820)
