+++ cvat-ui/src/actions/annotation-actions.ts
@@ -8,7 +8,7 @@ import {
     ActionCreator,
     Store,
 } from 'redux';
-import { ThunkAction } from 'redux-thunk';
+import { ThunkAction } from 'utils/redux';
 
 import {
     CombinedState,
@@ -191,8 +191,7 @@ export enum AnnotationActionTypes {
     SAVE_LOGS_FAILED = 'SAVE_LOGS_FAILED',
 }
 
-export function saveLogsAsync():
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function saveLogsAsync(): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>) => {
         try {
             await logger.save();
@@ -236,7 +235,7 @@ export function switchZLayer(cur: number): AnyAction {
     };
 }
 
-export function fetchAnnotationsAsync(): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function fetchAnnotationsAsync(): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const {
@@ -308,8 +307,7 @@ export function updateCanvasContextMenu(
     };
 }
 
-export function removeAnnotationsAsync(sessionInstance: any):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function removeAnnotationsAsync(sessionInstance: any): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             await sessionInstance.annotations.clear();
@@ -333,8 +331,7 @@ ThunkAction<Promise<void>, {}, {}, AnyAction> {
     };
 }
 
-export function uploadJobAnnotationsAsync(job: any, loader: any, file: File):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function uploadJobAnnotationsAsync(job: any, loader: any, file: File): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const state: CombinedState = getStore().getState();
@@ -404,8 +401,7 @@ ThunkAction<Promise<void>, {}, {}, AnyAction> {
     };
 }
 
-export function changeJobStatusAsync(jobInstance: any, status: string):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function changeJobStatusAsync(jobInstance: any, status: string): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const oldStatus = jobInstance.status;
         try {
@@ -435,8 +431,7 @@ ThunkAction<Promise<void>, {}, {}, AnyAction> {
     };
 }
 
-export function collectStatisticsAsync(sessionInstance: any):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function collectStatisticsAsync(sessionInstance: any): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             dispatch({
@@ -477,7 +472,7 @@ export function propagateObjectAsync(
     objectState: any,
     from: number,
     to: number,
-): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const copy = {
@@ -542,7 +537,7 @@ export function changePropagateFrames(frames: number): AnyAction {
 }
 
 export function removeObjectAsync(sessionInstance: any, objectState: any, force: boolean):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             await sessionInstance.logger.log(LogType.deleteObject, { count: 1 });
@@ -659,7 +654,7 @@ export function switchPlay(playing: boolean): AnyAction {
 }
 
 export function changeFrameAsync(toFrame: number, fillBuffer?: boolean, frameStep?: number):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const state: CombinedState = getStore().getState();
         const { instance: job } = state.annotation.job;
@@ -751,7 +746,7 @@ ThunkAction<Promise<void>, {}, {}, AnyAction> {
 }
 
 export function undoActionAsync(sessionInstance: any, frame: number):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const state = getStore().getState();
@@ -794,7 +789,7 @@ ThunkAction<Promise<void>, {}, {}, AnyAction> {
 }
 
 export function redoActionAsync(sessionInstance: any, frame: number):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const state = getStore().getState();
@@ -906,7 +901,7 @@ export function confirmCanvasReady(): AnyAction {
     };
 }
 
-export function closeJob(): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function closeJob(): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const { jobInstance } = receiveAnnotationsParameters();
         if (jobInstance) {
@@ -924,7 +919,7 @@ export function getJobAsync(
     jid: number,
     initialFrame: number,
     initialFilters: string[],
-): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const state: CombinedState = getStore().getState();
@@ -1001,7 +996,7 @@ export function getJobAsync(
 }
 
 export function saveAnnotationsAsync(sessionInstance: any):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const { filters, frame, showAllInterpolationTracks } = receiveAnnotationsParameters();
 
@@ -1118,8 +1113,7 @@ export function splitTrack(enabled: boolean): AnyAction {
     };
 }
 
-export function updateAnnotationsAsync(statesToUpdate: any[]):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function updateAnnotationsAsync(statesToUpdate: any[]): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const {
             jobInstance,
@@ -1164,7 +1158,7 @@ ThunkAction<Promise<void>, {}, {}, AnyAction> {
 }
 
 export function createAnnotationsAsync(sessionInstance: any, frame: number, statesToCreate: any[]):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const { filters, showAllInterpolationTracks } = receiveAnnotationsParameters();
@@ -1192,7 +1186,7 @@ ThunkAction<Promise<void>, {}, {}, AnyAction> {
 }
 
 export function mergeAnnotationsAsync(sessionInstance: any, frame: number, statesToMerge: any[]):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const { filters, showAllInterpolationTracks } = receiveAnnotationsParameters();
@@ -1230,7 +1224,7 @@ export function groupAnnotationsAsync(
     sessionInstance: any,
     frame: number,
     statesToGroup: any[],
-): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const { filters, showAllInterpolationTracks } = receiveAnnotationsParameters();
@@ -1266,7 +1260,7 @@ export function groupAnnotationsAsync(
 }
 
 export function splitAnnotationsAsync(sessionInstance: any, frame: number, stateToSplit: any):
-ThunkAction<Promise<void>, {}, {}, AnyAction> {
+ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const { filters, showAllInterpolationTracks } = receiveAnnotationsParameters();
         try {
@@ -1296,7 +1290,7 @@ ThunkAction<Promise<void>, {}, {}, AnyAction> {
 export function changeLabelColorAsync(
     label: any,
     color: string,
-): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const {
@@ -1334,7 +1328,7 @@ export function changeLabelColorAsync(
 export function changeGroupColorAsync(
     group: number,
     color: string,
-): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const state: CombinedState = getStore().getState();
         const groupStates = state.annotation.annotations.states
@@ -1352,7 +1346,7 @@ export function searchAnnotationsAsync(
     sessionInstance: any,
     frameFrom: number,
     frameTo: number,
-): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         try {
             const { filters } = receiveAnnotationsParameters();
@@ -1371,7 +1365,7 @@ export function searchAnnotationsAsync(
     };
 }
 
-export function pasteShapeAsync(): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function pasteShapeAsync(): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const {
             canvas: {
@@ -1430,7 +1424,7 @@ export function pasteShapeAsync(): ThunkAction<Promise<void>, {}, {}, AnyAction>
     };
 }
 
-export function repeatDrawShapeAsync(): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function repeatDrawShapeAsync(): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const {
             canvas: {
@@ -1494,7 +1488,7 @@ export function repeatDrawShapeAsync(): ThunkAction<Promise<void>, {}, {}, AnyAc
     };
 }
 
-export function redrawShapeAsync(): ThunkAction<Promise<void>, {}, {}, AnyAction> {
+export function redrawShapeAsync(): ThunkAction {
     return async (dispatch: ActionCreator<Dispatch>): Promise<void> => {
         const {
             annotations: {
