+++ cvat-ui/src/components/annotation-page/attribute-annotation-workspace/attribute-annotation-sidebar/object-basics-edtior.tsx
@@ -4,24 +4,15 @@
 
 import React from 'react';
 import Select, { SelectValue } from 'antd/lib/select';
-import Checkbox, { CheckboxChangeEvent } from 'antd/lib/checkbox';
 
 interface Props {
     currentLabel: string;
     labels: any[];
-    occluded: boolean;
-    setOccluded(event: CheckboxChangeEvent): void;
     changeLabel(value: SelectValue): void;
 }
 
 function ObjectBasicsEditor(props: Props): JSX.Element {
-    const {
-        currentLabel,
-        occluded,
-        labels,
-        setOccluded,
-        changeLabel,
-    } = props;
+    const { currentLabel, labels, changeLabel } = props;
 
     return (
         <div className='attribute-annotation-sidebar-basics-editor'>
@@ -35,7 +26,6 @@ function ObjectBasicsEditor(props: Props): JSX.Element {
                     </Select.Option>
                 ))}
             </Select>
-            <Checkbox checked={occluded} onChange={setOccluded}>Occluded</Checkbox>
         </div>
     );
 }
